feat: Add OAuth2 and OIDC utility types and functions
- Introduced OAuth2ClientConfig, PKCECodes, IdTokenPayload, JWKSKey, JWKSResponse, and OIDCDiscoveryDocument interfaces in oauth2.ts for structured OAuth2 configurations and token handling.
- Created TokenResponse, UserInfo, and TokenStorage interfaces in token.ts to manage token responses and user information.
- Developed EnvironmentUtils class in environment.ts for environment detection, crypto handling, and JWT parsing.
- Implemented OIDCUtils class in oidc.ts for OpenID Connect functionalities, including fetching discovery documents, JWKS, and validating ID tokens with RSA signatures.
